# app/implementation/openapi_server/models/character.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model_ import Model
from openapi_server import util


class Character(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, id=None, media_list=None, name=None, actor=None, picture=None):  # noqa: E501
        """Character - a model defined in OpenAPI

        :param id: The id of this Character.  # noqa: E501
        :type id: str
        :param media_list: The media_list of this Character.  # noqa: E501
        :type media_list: List[StaffMediaList]
        :param name: The name of this Character.  # noqa: E501
        :type name: str
        :param actor: The actor of this Character.  # noqa: E501
        :type actor: CharacterActor
        :param picture: The picture of this Character.  # noqa: E501
        :type picture: str
        """
        self.openapi_types = {
            'id': str,
            'media_list': List[StaffMediaList],
            'name': str,
            'actor': CharacterActor,
            'picture': str
        }

        self.attribute_map = {
            'id': 'id',
            'media_list': 'media_list',
            'name': 'name',
            'actor': 'actor',
            'picture': 'picture'
        }

        self._id = id
        self._media_list = media_list
        self._name = name
        self._actor = actor
        self._picture = picture

    @classmethod
    def from_dict(cls, dikt) -> 'Character':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Character of this Character.  # noqa: E501
        :rtype: Character
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self):
        """Gets the id of this Character.


        :return: The id of this Character.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id):
        """Sets the id of this Character.


        :param id: The id of this Character.
        :type id: str
        """

        self._id = id

    @property
    def media_list(self):
        """Gets the media_list of this Character.


        :return: The media_list of this Character.
        :rtype: List[StaffMediaList]
        """
        return self._media_list

    @media_list.setter
    def media_list(self, media_list):
        """Sets the media_list of this Character.


        :param media_list: The media_list of this Character.
        :type media_list: List[StaffMediaList]
        """

        self._media_list = media_list

    @property
    def name(self):
        """Gets the name of this Character.


        :return: The name of this Character.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name):
        """Sets the name of this Character.


        :param name: The name of this Character.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name

    @property
    def actor(self):
        """Gets the actor of this Character.


        :return: The actor of this Character.
        :rtype: CharacterActor
        """
        return self._actor

    @actor.setter
    def actor(self, actor):
        """Sets the actor of this Character.


        :param actor: The actor of this Character.
        :type actor: CharacterActor
        """

        self._actor = actor

    @property
    def picture(self):
        """Gets the picture of this Character.


        :return: The picture of this Character.
        :rtype: str
        """
        return self._picture

    @picture.setter
    def picture(self, picture):
        """Sets the picture of this Character.


        :param picture: The picture of this Character.
        :type picture: str
        """

        self._picture = picture