# app/implementation/openapi_server/models/staff.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model_ import Model
from openapi_server import util


class Staff(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, id=None, media_list=None, name=None, picture=None):  # noqa: E501
        """Staff - a model defined in OpenAPI

        :param id: The id of this Staff.  # noqa: E501
        :type id: str
        :param media_list: The media_list of this Staff.  # noqa: E501
        :type media_list: List[StaffMediaList]
        :param name: The name of this Staff.  # noqa: E501
        :type name: str
        :param picture: The picture of this Staff.  # noqa: E501
        :type picture: str
        """
        self.openapi_types = {
            'id': str,
            'media_list': List[StaffMediaList],
            'name': str,
            'picture': str
        }

        self.attribute_map = {
            'id': 'id',
            'media_list': 'media_list',
            'name': 'name',
            'picture': 'picture'
        }

        self._id = id
        self._media_list = media_list
        self._name = name
        self._picture = picture

    @classmethod
    def from_dict(cls, dikt) -> 'Staff':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Staff of this Staff.  # noqa: E501
        :rtype: Staff
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self):
        """Gets the id of this Staff.


        :return: The id of this Staff.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id):
        """Sets the id of this Staff.


        :param id: The id of this Staff.
        :type id: str
        """

        self._id = id

    @property
    def media_list(self):
        """Gets the media_list of this Staff.


        :return: The media_list of this Staff.
        :rtype: List[StaffMediaList]
        """
        return self._media_list

    @media_list.setter
    def media_list(self, media_list):
        """Sets the media_list of this Staff.


        :param media_list: The media_list of this Staff.
        :type media_list: List[StaffMediaList]
        """

        self._media_list = media_list

    @property
    def name(self):
        """Gets the name of this Staff.


        :return: The name of this Staff.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name):
        """Sets the name of this Staff.


        :param name: The name of this Staff.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name

    @property
    def picture(self):
        """Gets the picture of this Staff.


        :return: The picture of this Staff.
        :rtype: str
        """
        return self._picture

    @picture.setter
    def picture(self, picture):
        """Sets the picture of this Staff.


        :param picture: The picture of this Staff.
        :type picture: str
        """

        self._picture = picture